env: 디렉토리 명 변경(Controller, Service), 디렉토리 변경으로 클래스 명 변경, Response 공통 / 각 디렉토리 추가
