+++ src/main/java/com/bookstore/readme/domain/member/controller/MemberController.java
@@ -1,10 +1,9 @@
-package com.bookstore.readme.domain.member.api;
+package com.bookstore.readme.domain.member.controller;
 
-import com.bookstore.readme.domain.member.application.MemberService;
-import com.bookstore.readme.domain.member.dto.MemberDto;
+import com.bookstore.readme.domain.member.service.MemberService;
 import com.bookstore.readme.domain.member.dto.MemberJoinDto;
 import com.bookstore.readme.domain.member.dto.MemberLoginDto;
-import com.bookstore.readme.domain.member.dto.MemberResponse;
+import com.bookstore.readme.domain.member.response.MemberResponse;
 import com.bookstore.readme.domain.member.exception.DuplicationMemberException;
 import com.bookstore.readme.domain.member.exception.MemberErrorResponse;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/member")
-public class MemberApi {
+public class MemberController {
 
     private final MemberService memberService;
