+++ src/main/java/com/bookstore/readme/domain/member/dto/MemberDto.java
@@ -1,6 +1,6 @@
 package com.bookstore.readme.domain.member.dto;
 
-import com.bookstore.readme.domain.member.domain.Member;
+import com.bookstore.readme.domain.member.model.Member;
 import lombok.*;
 
 @Getter
