+++ src/main/java/com/bookstore/readme/domain/member/dto/MemberLoginDto.java
@@ -1,6 +1,5 @@
 package com.bookstore.readme.domain.member.dto;
 
-import com.bookstore.readme.domain.member.domain.Member;
 import lombok.Data;
 
 @Data
