+++ src/main/java/com/bookstore/readme/domain/member/model/Member.java
@@ -1,4 +1,4 @@
-package com.bookstore.readme.domain.member.domain;
+package com.bookstore.readme.domain.member.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.Column;
@@ -6,8 +6,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import lombok.*;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
