+++ src/main/java/com/bookstore/readme/domain/member/repository/MemberRepository.java
@@ -1,7 +1,6 @@
 package com.bookstore.readme.domain.member.repository;
 
-import com.bookstore.readme.domain.member.domain.Member;
-import com.bookstore.readme.domain.member.dto.MemberDto;
+import com.bookstore.readme.domain.member.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
